refactor - rename folder
I don't like moving into the direction of naming folders as if building a clean architecture
